+++ js/main.js
@@ -1,1 +1,4 @@
-import './draw-pictures.js';
+import {render as renderPictures} from './render.js';
+import {descriptionPhotosList} from './data.js';
+
+renderPictures(descriptionPhotosList);
